requirements.txt
-e .
aws-cdk.aws-cloudwatch~=1.110.0
aws-cdk.aws-codebuild~=1.110.0
aws-cdk.aws-codecommit~=1.110.0
aws-cdk.aws-codedeploy~=1.110.0
aws-cdk.aws-codepipeline~=1.110.0
aws-cdk.aws-codepipeline-actions~=1.110.0
aws-cdk.aws-dynamodb~=1.110.0
aws-cdk.aws-ec2~=1.110.0
aws-cdk.aws-glue~=1.110.0
aws-cdk.aws-iam~=1.110.0
aws-cdk.aws-kms~=1.110.0
aws-cdk.aws-lambda~=1.110.0
aws-cdk.aws-lambda-event-sources~=1.110.0
aws-cdk.aws-s3~=1.110.0
aws-cdk.aws-s3-assets~=1.110.0
aws-cdk.aws-s3-deployment~=1.110.0
aws-cdk.aws-s3-notifications~=1.110.0
aws-cdk.aws-secretsmanager~=1.110.0
aws-cdk.aws-sns~=1.110.0
aws-cdk.aws-sns-subscriptions~=1.110.0
aws-cdk.aws-ssm~=1.110.0
aws-cdk.aws-stepfunctions~=1.110.0
aws-cdk.aws-stepfunctions-tasks~=1.110.0
aws-cdk.core~=1.110.0
aws-cdk.lambda-layer-awscli~=1.110.0
aws-cdk.pipelines~=1.110.0
boto3~=1.17.97
